kernel/amd64: lapic: Fix icr_lo overflow bug
Somehow icr_lo was a uint8_t which is obviously not right and would
cause overflow bugs.

+++ sys/arch/amd64/amd64/lapic.c
@@ -273,7 +273,7 @@ void
 lapic_send_ipi(uint8_t id, uint8_t shorthand, uint8_t vector)
 {
     const uint32_t x2APIC_IPI_SELF = 0x3F0;
-    uint8_t icr_lo = vector | IPI_DEST_PHYSICAL;
+    uint64_t icr_lo = vector | IPI_DEST_PHYSICAL;
     bool x2apic_supported = has_x2apic();
 
     if (x2apic_supported && shorthand == IPI_SHORTHAND_SELF) {
